Relocate snapshot tests
